requirements.txt
flask_mongoengine==0.9.5
Werkzeug==1.0.1
Flask==1.1.2
autocorrect==2.2.2
datefinder==0.7.1
flask_cors==3.0.9
gensim==3.8.3
Image==1.5.33
imutils==0.5.3
joblib==0.17.0
keras==2.4.3
lxml==4.6.1
matplotlib==3.3.2
nltk==3.5
numpy==1.19.3
openpyxl==3.0.5
pandas==1.1.3
pdf2image==1.14.0
Pillow==8.0.1
pytesseract==0.3.6
python_dateutil==2.8.1
scikit_learn==0.23.2
scipy==1.4.1
tqdm==4.51.0
xlsxwriter==1.3.7
